feat: use _ejs to identify template file (#451)
* feat: use _ejs to identify template file

* fix: update ejsRender

* chore: match more accurate
