+++ packages/generate-material/src/index.ts
@@ -6,11 +6,34 @@ import { getNpmTarball, getAndExtractTarball } from 'ice-npm-utils';
 
 import ejsRenderDir from './ejsRenderDir';
 import formatProject from './formatProject';
-import { IOptions, ITemplateOptions } from './type';
 
 /**
  * init component by template
  */
+export interface ITemplateOptions {
+  npmName: string; // @icedesign/ice-label
+  name?: string; // ice-label (english and variable)
+  kebabCaseName?: string; // ice-label
+  npmScope?: string; // @icedesign
+  title?: string; //
+  description?: string;
+  className?: string;
+  version?: string;
+  category?: string;
+  // web, miniapp...
+  projectTargets?: string[];
+  adaptor?: boolean;
+}
+
+export interface IOptions {
+  rootDir: string;
+  materialTemplateDir: string;
+  templateOptions: ITemplateOptions;
+  enablePegasus?: boolean;
+  enableDefPublish?: boolean;
+  materialType: 'component' | 'block' | 'scaffold';
+}
+
 export async function generateMaterial({
   rootDir,
   materialTemplateDir,
